SmartEMR_NLP/NLPWebservice/requirements.txt
absl-py==0.11.0
astunparse==1.6.3
awscli==1.18.168
blis==0.4.1
boto3==1.16.8
botocore==1.19.8
cachetools==4.2.0
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
cheroot==8.4.5
CherryPy==18.6.0
click==7.1.2
colorama==0.4.3
conllu==4.2.1
cryptography==3.3.1
cymem==2.0.3
dataclasses==0.6
dill==0.3.2
docutils==0.15.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.5/en_core_web_sm-2.2.5.tar.gz
Flask==1.1.2
Flask-Cors==3.0.10
flatbuffers==1.12
future==0.18.2
gast==0.3.3
gensim==3.8.0
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
itsdangerous==1.1.0
jaraco.classes==3.1.0
jaraco.collections==3.0.0
jaraco.functools==3.0.1
jaraco.text==3.2.0
jieba==0.42.1
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.17.0
Keras-Preprocessing==1.1.2
ln2sql==0.1
Markdown==3.3.3
MarkupSafe==1.1.1
medacy @ git+https://github.com/NLPatVCU/medaCy.git@e53b048150aee7841f6fa620a9f0998d46b5b3ca
more-itertools==8.6.0
msgpack==0.5.6
msgpack-numpy==0.4.3.2
multiprocess==0.70.10
murmurhash==1.0.2
mysql-connector==2.2.9
nmslib==2.0.6
numpy==1.19.3
oauthlib==3.1.0
opt-einsum==3.3.0
panda==0.3.1
pandas==1.1.4
pathos==0.2.6
plac==1.1.3
portend==2.6
pox==0.2.8
ppft==1.6.6.2
preshed==3.0.2
protobuf==3.14.0
psutil==5.7.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.6.0
pycparser==2.20
python-crfsuite==0.9.7
python-dateutil==2.8.1
pytorch-crf==0.7.2
pytz==2020.1
PyYAML==5.3.1
regex==2020.10.28
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.5
s3transfer==0.3.3
sacremoses==0.0.43
scikit-learn==0.23.2
scipy==1.5.3
scispacy==0.2.2
sentencepiece==0.1.94
six==1.15.0
sklearn-crfsuite==0.3.6
smart-open==3.0.0
spacy==2.2.2
srsly==1.0.2
tabulate==0.8.7
tempora==4.0.0
termcolor==1.1.0
thinc==7.3.1
threadpoolctl==2.1.0
torch==1.7.1
tqdm==4.51.0
transformers==2.3.0
typing-extensions==3.7.4.3
ujson==4.0.1
urllib3==1.25.11
wasabi==0.8.0
Werkzeug==1.0.1
wrapt==1.12.1
xmltodict==0.12.0
zc.lockfile==2.0
